+++ packages/xchain-avax/src/index.ts
@@ -1,5 +1,10 @@
 // Export all elements from the 'client' module
-export * from './client'
+import { ClientKeystore } from './ClientKeystore'
+
+export { ClientKeystore, ClientKeystore as Client } from './ClientKeystore'
+export { ClientLedger } from './ClientLedger'
 
 // Export all elements from the 'const' module
 export * from './const'
+
+export default ClientKeystore