+++ packages/xchain-evm/src/types/client-types.ts
@@ -1,5 +1,14 @@
-import { FeeOption, Fees, Network } from '@xchainjs/xchain-client'
-import { Address, BaseAmount } from '@xchainjs/xchain-util'
+import { Provider } from '@ethersproject/abstract-provider'
+import {
+  EvmOnlineDataProviders,
+  ExplorerProviders,
+  FeeOption,
+  Fees,
+  Network,
+  TxParams as BaseTxParams,
+  XChainClientParams,
+} from '@xchainjs/xchain-client'
+import { Address, Asset, BaseAmount, Chain } from '@xchainjs/xchain-util'
 import { BigNumber, Signer, ethers } from 'ethers'
 
 export type ClientUrl = Record<Network, string>
@@ -21,7 +30,6 @@ export type FeesWithGasPricesAndLimits = { fees: Fees; gasPrices: GasPrices; gas
 
 export type ApproveParams = {
   walletIndex?: number
-  signer?: Signer
   contractAddress: Address
   spenderAddress: Address
   feeOption?: FeeOption
@@ -54,3 +62,31 @@ export type CallParams = {
 }
 
 export type EstimateCallParams = Pick<CallParams, 'contractAddress' | 'abi' | 'funcName' | 'funcParams'>
+
+export type EvmDefaults = {
+  transferGasAssetGasLimit: BigNumber
+  transferTokenGasLimit: BigNumber
+  approveGasLimit: BigNumber
+  gasPrice: BigNumber // BaseAmount Unit
+}
+/**
+ * Parameters for configuring the EVM client.
+ */
+export type EVMClientParams = XChainClientParams & {
+  chain: Chain
+  gasAsset: Asset
+  gasAssetDecimals: number
+  defaults: Record<Network, EvmDefaults>
+  providers: Record<Network, Provider>
+  explorerProviders: ExplorerProviders
+  dataProviders: EvmOnlineDataProviders[]
+}
+
+export type TxParams = BaseTxParams & {
+  feeOption?: FeeOption
+  gasPrice?: BaseAmount
+  maxFeePerGas?: BaseAmount
+  maxPriorityFeePerGas?: BaseAmount
+  gasLimit?: BigNumber
+  isMemoEncoded?: boolean
+}