+++ examples/do-swap/CHANGELOG.md
@@ -1,5 +1,55 @@
 # xchainjs-do-swap
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [f432295]
+- Updated dependencies [448c29f]
+- Updated dependencies [f432295]
+- Updated dependencies [448c29f]
+- Updated dependencies [448c29f]
+- Updated dependencies [448c29f]
+- Updated dependencies [448c29f]
+- Updated dependencies [448c29f]
+- Updated dependencies [f432295]
+- Updated dependencies [f432295]
+  - @xchainjs/xchain-ethereum@0.32.0
+  - @xchainjs/xchain-avax@0.5.0
+  - @xchainjs/xchain-bsc@0.5.0
+  - @xchainjs/xchain-thorchain-amm@1.1.4
+  - @xchainjs/xchain-evm@0.5.0
+  - @xchainjs/xchain-wallet@0.1.9
+  - @xchainjs/xchain-client@0.16.3
+  - @xchainjs/xchain-binance@5.7.12
+  - @xchainjs/xchain-bitcoin@0.23.14
+  - @xchainjs/xchain-bitcoincash@0.17.12
+  - @xchainjs/xchain-cosmos@1.0.2
+  - @xchainjs/xchain-cosmos-sdk@0.2.5
+  - @xchainjs/xchain-dash@0.3.1
+  - @xchainjs/xchain-doge@0.7.14
+  - @xchainjs/xchain-evm-providers@0.1.8
+  - @xchainjs/xchain-litecoin@0.14.1
+  - @xchainjs/xchain-midgard-query@0.1.14
+  - @xchainjs/xchain-thorchain@1.0.5
+  - @xchainjs/xchain-thorchain-query@0.7.9
+  - @xchainjs/xchain-utxo@0.1.5
+  - @xchainjs/xchain-utxo-providers@0.2.13
+
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [8e82dc3]
+- Updated dependencies [8022a60]
+- Updated dependencies [8e82dc3]
+- Updated dependencies [2f7988a]
+  - @xchainjs/xchain-doge@0.7.13
+  - @xchainjs/xchain-dash@0.3.0
+  - @xchainjs/xchain-bitcoincash@0.17.11
+  - @xchainjs/xchain-litecoin@0.14.0
+  - @xchainjs/xchain-thorchain-amm@1.1.3
+
 ## 1.0.3
 
 ### Patch Changes