+++ packages/xchain-bitcoin/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.23.14
+
+### Patch Changes
+
+- Updated dependencies [448c29f]
+  - @xchainjs/xchain-client@0.16.3
+  - @xchainjs/xchain-utxo@0.1.5
+  - @xchainjs/xchain-utxo-providers@0.2.13
+
 ## 0.23.13
 
 ### Patch Changes