+++ packages/xchain-evm/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.5.0
+
+### Minor Changes
+
+- f432295: Ledger client
+
+### Patch Changes
+
+- 448c29f: New optional parameter `isMemoEncoded` for `transfer`, `prepareTx` and `estimateGasLimit` methods. If it is set to true, memo will not be encoded by the funcions. False by default
+- f432295: `approve` return type updated from `ethers.providers.TransactionResponse` to `string`
+- Updated dependencies [448c29f]
+  - @xchainjs/xchain-client@0.16.3
+
 ## 0.4.5
 
 ### Patch Changes