+++ packages/xchain-litecoin/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.14.1
+
+### Patch Changes
+
+- Updated dependencies [448c29f]
+  - @xchainjs/xchain-client@0.16.3
+  - @xchainjs/xchain-utxo@0.1.5
+  - @xchainjs/xchain-utxo-providers@0.2.13
+
+## 0.14.0
+
+### Minor Changes
+
+- 2f7988a: Ledger client
+
 ## 0.13.11
 
 ### Patch Changes