+++ packages/xchain-mayachain-amm/CHANGELOG.md
@@ -1,5 +1,45 @@
 # Changelog
 
+## 2.0.0
+
+### Patch Changes
+
+- 448c29f: EVM protocol actions through the router smart contract
+- Updated dependencies [f432295]
+- Updated dependencies [f432295]
+- Updated dependencies [448c29f]
+- Updated dependencies [448c29f]
+- Updated dependencies [448c29f]
+- Updated dependencies [448c29f]
+- Updated dependencies [448c29f]
+- Updated dependencies [f432295]
+- Updated dependencies [f432295]
+  - @xchainjs/xchain-ethereum@0.32.0
+  - @xchainjs/xchain-evm@0.5.0
+  - @xchainjs/xchain-wallet@0.1.9
+  - @xchainjs/xchain-client@0.16.3
+  - @xchainjs/xchain-bitcoin@0.23.14
+  - @xchainjs/xchain-dash@0.3.1
+  - @xchainjs/xchain-kujira@0.1.13
+  - @xchainjs/xchain-mayachain@1.0.2
+  - @xchainjs/xchain-mayachain-query@0.1.11
+  - @xchainjs/xchain-thorchain@1.0.5
+
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [ecacc18]
+- Updated dependencies [e0d9669]
+  - @xchainjs/xchain-mayachain-query@0.1.10
+
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [8022a60]
+  - @xchainjs/xchain-dash@0.3.0
+
 ## 1.1.0
 
 ### Minor Changes