+++ packages/xchain-mayamidgard-query/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies [448c29f]
+  - @xchainjs/xchain-client@0.16.3
+
+## 0.1.8
+
+### Patch Changes
+
+- e0d9669: Migard API backup endpoint
+
 ## 0.1.7
 
 ### Patch Changes